pyproject.toml
[project]
name = "seller-audit-team"
version = "0.1.0"
description = "PalmStreet seller audit workspace scripts and tooling."
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-cloud-bigquery",
    "pyyaml",
]

[tool.uv]
package = false
